# poslappapi/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.2"
  title: PosLapp
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /status:
    # binds a127 app logic to a route
    x-swagger-router-controller: status.controller
    get:
      description: get the equipo system list
      operationId: getStatus
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetStatusListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: add a new game system to the list
      operationId: createStatus
      parameters:
        - in: body
          name: status
          description: The Game System to create.
          schema:
            $ref: "#/definitions/Status"
      responses:
        "201":
          description: Success
          schema:
            $ref: "#/definitions/GetStatusResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /status/{id}:
    x-swagger-router-controller: status.controller
    get:
      description: get a Status
      operationId : getStatusById
      parameters:
        - name: id
          type: number
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetStatusResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    
    put:
      description: update a status
      operationId : updateStatus
      # define the parameters
      parameters:
        - name: id
          description: status id
          type: number
          in: path
          required: true
        - in: body
          name: statusUpdate
          description: The Status to update.
          schema:
            $ref: "#/definitions/Status"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetStatusResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  
    delete:
      description: delete a status  
      operationId: deleteStatus
      # define the parameters
      parameters:
        - name: id
          description: status id
          type: number
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  Status:
    type: object
    properties:
      id_estatus:
        type: integer
        description: Numerical value
      descripcionEstatus:
        type: string
        description: Name of status
    required:
      - id_estatus
      - descripcionEstatus
  GetStatusResponse:
    type: object
    properties:
      #id:
      #  type: number
      #  description: Numerical value
      id_estatus:
        type: number
        description: Numerical value
      descripcionEstatus:
        type: string
        description: Name of status
    required:
      - id_estatus
      - descripcionEstatus
  GetStatusListResponse:
    required:
      - statusList
    properties:
      statusList:
        type: array
        items:
          $ref: "#/definitions/GetStatusResponse"
  GeneralResponse:
    type: object
    properties:
      success:
        type: number
        description: returns 1 if successful
      description:
        type: string
        description: a short comment
    required:
      - success
      - description
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string